# common/trinary: trits and trytes conversions optimizations