100000000 iterations
trits_to_trytes before  135222 ms
trits_to_trytes after   117607 ms (-13.027%)

100000000 iterations
trytes_to_trits before  187789 ms
trytes_to_trits after   113821 ms (-39.389%)

# Test Plan:
CI